Adding double fault interrupt handler
Double fault interrupt can be triggered by many factor. In case of a stack overflow, the system no longer has a valid stack pointer to continue operation thus making the system bootlooping. To prevent this, we register a GlobalDefinitionsTable and keep track of healthy stack pointer of which out double_fault handler can refer to when double fault interrupt occurs.