Added heap allocator test many_boxes_long_lived
This test checks if the allocator is able to free previously freed block
of memory. Bump allocator will not be able to reuse previous memory block
unless you free all blocks and reset the head.
Our current implementation is a linked list allocator which keeps track
of unused block thus we are able to reuse freed blocks.